♻️  refactor: move TextArea component to components/textarea folder (#88)
